+++ apps/desktop/src-tauri/src/orchestrate.rs
@@ -949,6 +949,8 @@ mod tests {
             last_error: None,
             etag: None,
             last_modified: None,
+            userinfo: None,
+            provider_info: vec![],
             auto_update: false,
             auto_update_interval: None,
         };
@@ -1513,6 +1515,8 @@ mod tests {
             last_error: None,
             etag: None,
             last_modified: None,
+            userinfo: None,
+            provider_info: vec![],
             auto_update: false,
             auto_update_interval: None,
         };
@@ -1812,6 +1816,8 @@ mod tests {
             last_error: None,
             etag: None,
             last_modified: None,
+            userinfo: None,
+            provider_info: vec![],
             auto_update: false,
             auto_update_interval: None,
         };
@@ -1884,6 +1890,8 @@ mod tests {
             last_error: None,
             etag: None,
             last_modified: None,
+            userinfo: None,
+            provider_info: vec![],
             auto_update: false,
             auto_update_interval: None,
         };